% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionGraphs.R
\name{ig}
\alias{ig}
\title{Constructs Interaction Graph (S3 class)}
\usage{
ig(n, ne, pe)
}
\arguments{
\item{n}{ig.nodes \code{(a list of igNode-s)}}

\item{ne}{ig.negEdges \code{(a list of negative igEdge-s)}}

\item{pe}{ig.posEdges \code{(a list of negative igEdge-s)}}
}
\value{
An instance of the \code{ig} class
}
\description{
Constructs Interaction Graph (S3 class)
}
