% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionGraphs.R
\name{igEdge}
\alias{igEdge}
\title{Constructs Interaction Graph Edges (S3 class)}
\usage{
igEdge(n1, n2, w)
}
\arguments{
\item{n1}{igEdge.node1 \code{(character)}}

\item{n2}{igEdge.node2 \code{(character)}}

\item{w}{igEdge.weight \code{(double)} (i.e. 3-way Interaction Gain)}
}
\value{
An instance of the \code{igEdge} class
}
\description{
Constructs Interaction Graph Edges (S3 class)
}
