% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionGraphs.R
\name{igNode}
\alias{igNode}
\title{Constructs Interaction Graph Nodes (S3 class)}
\usage{
igNode(n, v)
}
\arguments{
\item{n}{igNode.name \code{(character)}}

\item{v}{igNode.value \code{(double)} (i.e. 2-way Interaction Gain)}
}
\value{
An instance of the \code{igNode} class
}
\description{
Constructs Interaction Graph Nodes (S3 class)
}
