% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpingFunctions.R
\name{InfoGain}
\alias{InfoGain}
\title{Calculates Information gain}
\usage{
InfoGain(df, inAtt, classAtt)
}
\arguments{
\item{df}{A dataframe (discrete!)}

\item{inAtt}{An input column of the dataframe}

\item{classAtt}{A class column of the dataframe}
}
\value{
the Shannon's entropy of \code{df} based on \code{classAtt}
}
\description{
Calculates Information gain
}
