% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionGraphs.R
\name{interactionGraph}
\alias{interactionGraph}
\title{Creates Interaction graph}
\usage{
interactionGraph(df, classAtt, intNo = 16, precomputed = FALSE)
}
\arguments{
\item{df}{A (discrete) data.frame}

\item{classAtt}{A class column of the df (string)}

\item{intNo}{A desired number of interactions (integer) to show. In range:
[2,20] in steps of 2; Default value is 16 (8 positive and 8 negative
interactions). If the provided value is odd, it will be rounded to the
closest largest value; if new valu is outside the range, it will be scaled
to fit (e.g. 1->2, 22->20).}

\item{precomputed}{A (boolean) indication if the \code{df} is pointing to a
data.frame with precomputed interaction gains (generated by
\code{interactions3Way} function) or to the original dataset. By default a
data.frame containing the original data set is expected
(\code{precomputed=FALSE}.}
}
\value{
An interaction graph object
}
\description{
Descirpiton...
}
