% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionGraphs.R
\name{interactions3Way}
\alias{interactions3Way}
\title{Calculates 3-Way Interactions}
\usage{
interactions3Way(df, classAtt, speedUp = FALSE)
}
\arguments{
\item{df}{A (discrete) data.frame}

\item{classAtt}{A class column of the df (string)}

\item{speedUp}{A (boolean) parameter, if TRUE, indicating whether pairs of
attributes with information gain equal to zero (on the 4th decimal) should
be pruned. This speeds up calculations for larger datasets. By default it
is turned off (i.e. set to FALSE).}
}
\value{
A data-frame with 3-way interactions
}
\description{
Formula: \eqn{}, where \eqn{} is
}
