% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpingFunctions.R
\name{IsDiscreteDataFrame}
\alias{IsDiscreteDataFrame}
\title{Tests if all columns of the data.frame are discrete (i.e. factors)}
\usage{
IsDiscreteDataFrame(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
Boolean: TRUE if \code{obj} all columns of the data frame \code{df}
are factors, FALSE otherwise
}
\description{
Tests if all columns of the data.frame are discrete (i.e. factors)
}
