% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunctions.R
\name{plotIntGraph}
\alias{plotIntGraph}
\title{Plots Interaction graph}
\usage{
plotIntGraph(ig)
}
\arguments{
\item{ig}{Interaction graph}
}
\value{
Plots the \code{ig}
}
\description{
Plots Interaction graph
}
\examples{
plotIntGraph(interactionGraph(golf, "Play", intNo = 10))
}
