% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionGraphs.R
\name{print.ig}
\alias{print.ig}
\title{Print generic method for Interaction Graph (S3 class)}
\usage{
\method{print}{ig}(intGraph)
}
\arguments{
\item{intGraph}{An \code{(ig)} object}
}
\value{
Print \code{(ig)} object
}
\description{
Print generic method for Interaction Graph (S3 class)
}
