% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionGraphs.R
\name{print.igEdge}
\alias{print.igEdge}
\title{Print generic method for Interaction Graph Edges (S3 class)}
\usage{
\method{print}{igEdge}(edge)
}
\arguments{
\item{edge}{An \code{(igEdge)} object}
}
\value{
Print \code{(igEdge)} object
}
\description{
Print generic method for Interaction Graph Edges (S3 class)
}
