% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionGraphs.R
\name{print.igNode}
\alias{print.igNode}
\title{Print generic method for Interaction Graph Nodes (S3 class)}
\usage{
\method{print}{igNode}(node)
}
\arguments{
\item{node}{An \code{(igNode)} object}
}
\value{
Print \code{(igNode)} object
}
\description{
Print generic method for Interaction Graph Nodes (S3 class)
}
