% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionGraphs.R
\name{toString.ig}
\alias{toString.ig}
\title{toString() generic method for Interaction Graph (S3 class)}
\usage{
\method{toString}{ig}(intGraph)
}
\arguments{
\item{intGraph}{An \code{ig} object}
}
\value{
A \code{character} object made of the provided \code{ig} object
}
\description{
toString() generic method for Interaction Graph (S3 class)
}
