% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionGraphs.R
\name{toString.igEdge}
\alias{toString.igEdge}
\title{toString() generic method for Interaction Graph Edges (S3 class)}
\usage{
\method{toString}{igEdge}(edge)
}
\arguments{
\item{edge}{An \code{(igEdge)} object}
}
\value{
\code{(character)} object made of the provided \code{(igEdge)} object
}
\description{
toString() generic method for Interaction Graph Edges (S3 class)
}
