% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionGraphs.R
\name{toString.igNode}
\alias{toString.igNode}
\title{toString() generic method for Interaction Graph Nodes (S3 class)}
\usage{
\method{toString}{igNode}(node)
}
\arguments{
\item{node}{An \code{(igNode)} object}
}
\value{
\code{(character)} object made of the provided \code{(igNode)} object
}
\description{
toString() generic method for Interaction Graph Nodes (S3 class)
}
