% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpingFunctions.R
\name{Entropy}
\alias{Entropy}
\title{Calculates Shannon's entropy}
\usage{
Entropy(df, classAtt)
}
\arguments{
\item{df}{A dataframe (discrete!) (name, string)}

\item{classAtt}{A class column of the df (name, string)}
}
\value{
the Shannon's entropy of \code{df} based on \code{classAtt}
}
\description{
Calculates Shannon's entropy
}
