% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionGraphs.R
\name{ig}
\alias{ig}
\title{Constructs Interaction Graph (S3 class)}
\usage{
ig(n, e)
}
\arguments{
\item{n}{ig.nodes (\code{a list of igNode objects})}

\item{e}{ig.edges (\code{a list of igEdge objects})}
}
\value{
An instance of the \code{ig} class
}
\description{
Constructs Interaction Graph (S3 class)
}
