% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunctions.R
\name{igToPDF}
\alias{igToPDF}
\title{Exports Interaction graph to a PDF file}
\usage{
igToPDF(ig, fName = "InteractionGraph", path = getwd(), h = 2000)
}
\arguments{
\item{ig}{Interaction graph}

\item{fName}{The name of the file to be created; "InteractionGraph" by
default}

\item{path}{The folder in which to write the PDF file; by default,
the working directory}

\item{h}{Desired height of the image in pixels; 2000px by default}
}
\value{
Writes the \code{ig} interaction graph to a PDF (\code{.pdf}) file
  to the specified folder in \code{path}
}
\description{
Exports Interaction graph to a PDF file
}
