% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunctions.R
\name{igToPNG}
\alias{igToPNG}
\title{Exports Interaction graph to a PNG file}
\usage{
igToPNG(ig, fName = "InteractionGraph", path = getwd(), h = 2000)
}
\arguments{
\item{ig}{Interaction graph}

\item{fName}{The name of the file to be created; "InteractionGraph" by
default}

\item{path}{The folder in which to write the PNG file; by default,
the working directory}

\item{h}{Desired height of the image in pixels; 2000px by default}
}
\value{
Writes the \code{ig} interaction graph to a PNG (\code{.png}) file
  to the folder specified in the \code{path}
}
\description{
Exports Interaction graph to a PNG file
}
\examples{
\dontrun{
g <- interactionGraph(golf, "Play", intNo = 10)
igToPNG(g)
}
\dontrun{
g <- interactionGraph(golf, "Play", intNo = 10)
igToPNG(g, path = "C:/myFolder", fName = "MyGraph", h = 2000)
}
}
