% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunctions.R
\name{igToPS}
\alias{igToPS}
\title{Exports Interaction graph to a PS (PostScript) file}
\usage{
igToPS(ig, fName = "InteractionGraph", path = getwd(), height = 2000)
}
\arguments{
\item{ig}{Interaction graph}

\item{fName}{The name of the file to be created; "InteractionGraph" by
default}

\item{path}{The folder in which to write the PS file; by default,
the working directory}

\item{height}{Desired height of the image in pixels; 2000px by default}
}
\value{
Writes the \code{ig} interaction graph to a PDF (\code{.ps}) file
  to the specified folder in \code{path}
}
\description{
Exports Interaction graph to a PS (PostScript) file
}
