% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFunctions.R
\name{igToSVG}
\alias{igToSVG}
\title{Exports Interaction graph to a SVG file}
\usage{
igToSVG(ig, path = getwd(), fName = "InteractionGraph",
  height = 2000)
}
\arguments{
\item{ig}{Interaction graph}

\item{path}{The folder in which to write the SVG file; by default,
the working directory}

\item{fName}{The name of the file to be created; "InteractionGraph" by
default}

\item{height}{Desired height of the image in pixels; 2000px by default}
}
\value{
Writes the \code{ig} interaction graph to a SVG (\code{.svg}) file
  to the specified folder in \code{path}
}
\description{
Exports Interaction graph to a SVG file
}
