% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionGraphs.R
\name{interactionGraph}
\alias{interactionGraph}
\title{Creates Interaction graph}
\usage{
interactionGraph(df, classAtt, intNo = 16, speedUp = FALSE)
}
\arguments{
\item{df}{A \code{discrete} \code{data.frame}}

\item{classAtt}{A class column of the df (\code{string})}

\item{intNo}{A desired number of interactions to show, i.e. an
(\code{integer}) in range: [2,20]; Default value is 16.}
}
\value{
An interaction graph object (\code{string})
}
\description{
Creates Interaction graph
}
\examples{
\dontrun{interactionGraph(golf, "Play")}
\dontrun{interactionGraph(golf, "Play", intNo = 10, speedUp = TRUE)}
}
