% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpingFunctions.R
\name{isDiscreteDataFrame}
\alias{isDiscreteDataFrame}
\title{Tests if data.frame is discrete (i.e. all of its columns are factors)}
\usage{
isDiscreteDataFrame(df)
}
\arguments{
\item{df}{A \code{data.frame}}
}
\value{
\code{Boolean}: \code{TRUE} if all columns of the \code{data.frame}
  \code{df} are factors, \code{FALSE} otherwise; If the provided \code{df}
  object is of other type than \code{data.frame}, the function throws an
  error.
}
\description{
Tests if data.frame is discrete (i.e. all of its columns are factors)
}
\examples{
\dontrun{isDiscreteDataFrame(golf)}
}
