% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpingFunctions.R
\name{IsDiscreteDataFrame}
\alias{IsDiscreteDataFrame}
\title{Tests if data.frame is discrete (i.e. all of its columns are factors)}
\usage{
IsDiscreteDataFrame(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
Boolean: TRUE if all columns of the data frame \code{df}
are factors, FALSE otherwise; If the provided \code{df} object is of other
type than data.frame, the function returns FALSE.
}
\description{
Tests if data.frame is discrete (i.e. all of its columns are factors)
}
